link technologies to assets
