+++ frontend/src/components/projects/assets/GenericAssetCreateDialog.vue
@@ -1,11 +1,12 @@
 <script>
 import AssetService from '@/service/AssetService';
-import AssetEnvironmentSelectField from '../elements/forms/AssetEnvironmentSelectField.vue';
-import AssetAccessibleSelectField from '../elements/forms/AssetAccessibleSelectField.vue';
+import AssetEnvironmentSelectField from '@/components/elements/forms/AssetEnvironmentSelectField.vue';
+import AssetAccessibleSelectField from '@/components/elements/forms/AssetAccessibleSelectField.vue';
 import MarkdownEditor from '@/components/forms/MarkdownEditor.vue';
+import TechnologyMultiSelectField from '@/components/forms/fields/TechnologyMultiSelectField.vue';
 
 export default {
-    name: 'HostCreateDialog',
+    name: 'GenericAssetCreateDialog',
     emits: ['object-created'],
     data() {
         return {
@@ -16,9 +17,10 @@ export default {
                 dns: null,
                 operating_system: null,
                 environment: 'Unknown',
-                accessible: 'Unknown'
+                accessible: 'Unknown',
+                technologies: []
             },
-            assetService: new AssetService()
+            service: new AssetService()
         };
     },
     methods: {
@@ -29,7 +31,7 @@ export default {
             this.visible = true;
         },
         create() {
-            this.assetService.createHost(this.$api, this.projectId, this.model).then((response) => {
+            this.service.createGenericAsset(this.$api, this.projectId, this.model).then((response) => {
                 this.$toast.add({
                     severity: 'success',
                     summary: 'Created!',
@@ -41,26 +43,18 @@ export default {
             });
         }
     },
-    components: { AssetEnvironmentSelectField, AssetAccessibleSelectField, MarkdownEditor }
+    components: { TechnologyMultiSelectField, AssetEnvironmentSelectField, AssetAccessibleSelectField, MarkdownEditor }
 };
 </script>
 
 <template>
-    <Button icon="fa fa-plus" label="Host" outlined @click="open"></Button>
+    <Button icon="fa fa-plus" label="Generic Asset" outlined @click="open"></Button>
 
-    <Dialog header="Create Host" v-model:visible="visible" modal :style="{ width: '70vw' }">
+    <Dialog header="Create Generic Asset" v-model:visible="visible" modal :style="{ width: '70vw' }">
         <div class="p-fluid formgrid grid">
             <div class="field col-12">
-                <label for="ip">IP</label>
-                <InputText id="ip" type="text" v-model="model.ip"></InputText>
-            </div>
-            <div class="field col-12 md:col-6">
-                <label for="dns">DNS</label>
-                <InputText id="dns" type="text" v-model="model.dns"></InputText>
-            </div>
-            <div class="field col-12 md:col-6">
-                <label for="os">Operating System</label>
-                <InputText id="os" type="text" v-model="model.operating_system"></InputText>
+                <label for="name">Name</label>
+                <InputText id="name" type="text" v-model="model.name"></InputText>
             </div>
 
             <div class="field col-12 md:col-6">
@@ -68,6 +62,10 @@ export default {
             </div>
             <div class="field col-12 md:col-6">
                 <AssetAccessibleSelectField v-model="model.accessible"></AssetAccessibleSelectField>
+            </div>
+            <div class="field col-12">
+                <label for="technologies">Technologies</label>
+                <TechnologyMultiSelectField v-model="model.technologies"></TechnologyMultiSelectField>
             </div>
             <div class="field col-12">
                 <label for="description">Description</label>
