+++ frontend/src/components/projects/assets/MobileApplicationUpdateDialog.vue
@@ -3,9 +3,10 @@ import AssetService from '@/service/AssetService';
 import AssetEnvironmentSelectField from '@/components/elements/forms/AssetEnvironmentSelectField.vue';
 import AssetAccessibleSelectField from '@/components/elements/forms/AssetAccessibleSelectField.vue';
 import MarkdownEditor from '@/components/forms/MarkdownEditor.vue';
+import TechnologyMultiSelectField from '@/components/forms/fields/TechnologyMultiSelectField.vue';
 
 export default {
-    name: 'GenericAssetUpdateDialog',
+    name: 'MobileApplicationUpdateDialog',
     props: {
         asset: {
             required: true
@@ -16,7 +17,21 @@ export default {
         return {
             visible: false,
             model: this.asset,
-            service: new AssetService()
+            service: new AssetService(),
+            osChoices: [
+                {
+                    label: 'Unknown',
+                    value: 'Unknown'
+                },
+                {
+                    label: 'Android',
+                    value: 'Android'
+                },
+                {
+                    label: 'iOS',
+                    value: 'iOS'
+                }
+            ]
         };
     },
     methods: {
@@ -29,11 +44,17 @@ export default {
         patch() {
             let data = {
                 name: this.model.name,
+                version: this.model.version,
                 environment: this.model.environment,
                 accessible: this.model.accessible,
-                description: this.model.description
+                description: this.model.description,
+                os: this.model.os,
+                technologies: this.model.technologies
             };
-            this.service.patchGenericAsset(this.$api, this.$route.params.projectId, this.asset.pk, data).then(() => {
+            if (this.model.technologies.length > 0 && this.model.technologies[0].pk) {
+                delete data.technologies;
+            }
+            this.service.patchMobileApplication(this.$api, this.$route.params.projectId, this.asset.pk, data).then(() => {
                 this.$emit('object-updated', this.model);
                 this.visible = false;
             });
@@ -48,25 +69,42 @@ export default {
             }
         }
     },
-    components: { AssetEnvironmentSelectField, AssetAccessibleSelectField, MarkdownEditor }
+    components: { TechnologyMultiSelectField, AssetEnvironmentSelectField, AssetAccessibleSelectField, MarkdownEditor }
 };
 </script>
 
 <template>
     <Button icon="fa fa-pen-to-square" size="small" @click="open" outlined label="Edit"></Button>
 
-    <Dialog header="Update Generic Asset" v-model:visible="visible" :modal="true" :style="{ width: '70vw' }">
-        <div class="p-fluid formgrid grid">
+    <Dialog header="Update Web Application" v-model:visible="visible" :modal="true" :style="{ width: '70vw' }">
+        <div class="formgrid grid p-fluid">
             <div class="field col-12">
                 <label for="name">Name</label>
                 <InputText id="name" type="text" v-model="model.name"></InputText>
             </div>
-
+            <div class="field col-12 md:col-6">
+                <label for="os">Operating System</label>
+                <Dropdown :options="osChoices" option-label="label" option-value="value" v-model="model.os"></Dropdown>
+            </div>
+            <div class="field col-12 md:col-6">
+                <label for="version">Version</label>
+                <InputText id="version" type="text" v-model="model.version"></InputText>
+            </div>
+            <div class="field col-12">
+                <div class="flex align-items-center">
+                    <Checkbox v-model="model.certificate_pinning" inputId="cert_pinning" :binary="true" />
+                    <label for="cert_pinning" class="ml-2"> Certificate Pinning?</label>
+                </div>
+            </div>
             <div class="field col-12 md:col-6">
                 <AssetEnvironmentSelectField v-model="model.environment"></AssetEnvironmentSelectField>
             </div>
             <div class="field col-12 md:col-6">
                 <AssetAccessibleSelectField v-model="model.accessible"></AssetAccessibleSelectField>
+            </div>
+            <div class="field col-12">
+                <label for="technologies">Technologies</label>
+                <TechnologyMultiSelectField v-model="model.technologies"></TechnologyMultiSelectField>
             </div>
             <div class="field col-12">
                 <label for="description">Description</label>
