+++ frontend/src/components/projects/assets/WebApplicationUpdateDialog.vue
@@ -1,0 +1,102 @@
+<script>
+import AssetService from '@/service/AssetService';
+import AssetEnvironmentSelectField from '@/components/elements/forms/AssetEnvironmentSelectField.vue';
+import AssetAccessibleSelectField from '@/components/elements/forms/AssetAccessibleSelectField.vue';
+import MarkdownEditor from '@/components/forms/MarkdownEditor.vue';
+import TechnologyMultiSelectField from '@/components/forms/fields/TechnologyMultiSelectField.vue';
+
+export default {
+    name: 'WebApplicationUpdateDialog',
+    props: {
+        asset: {
+            required: true
+        }
+    },
+    emits: ['object-updated'],
+    data() {
+        return {
+            visible: false,
+            model: this.asset,
+            service: new AssetService(),
+            loading: false
+        };
+    },
+    methods: {
+        close() {
+            this.visible = false;
+        },
+        open() {
+            this.visible = true;
+        },
+        patch() {
+            this.loading = true;
+            let data = {
+                name: this.model.name,
+                base_url: this.model.base_url,
+                environment: this.model.environment,
+                accessible: this.model.accessible,
+                description: this.model.description,
+                technologies: this.model.technologies
+            };
+            if (this.model.technologies.length > 0 && this.model.technologies[0].pk) {
+                delete data.technologies;
+            }
+            this.service
+                .patchWebApplication(this.$api, this.$route.params.projectId, this.asset.pk, data)
+                .then(() => {
+                    this.$emit('object-updated', this.model);
+                    this.visible = false;
+                })
+                .finally(() => {
+                    this.loading = false;
+                });
+        }
+    },
+    watch: {
+        asset: {
+            immediate: true,
+            deep: true,
+            handler(value) {
+                this.model = value;
+            }
+        }
+    },
+    components: { TechnologyMultiSelectField, AssetEnvironmentSelectField, AssetAccessibleSelectField, MarkdownEditor }
+};
+</script>
+
+<template>
+    <Button icon="fa fa-pen-to-square" size="small" @click="open" outlined label="Edit"></Button>
+
+    <Dialog header="Update Web Application" v-model:visible="visible" :modal="true" :style="{ width: '70vw' }">
+        <div class="formgrid grid p-fluid">
+            <div class="field col-12">
+                <label for="name">Name</label>
+                <InputText id="name" type="text" v-model="model.name"></InputText>
+            </div>
+            <div class="field col-12">
+                <label for="base_url">Base URL</label>
+                <InputText id="base_url" type="text" v-model="model.base_url"></InputText>
+            </div>
+            <div class="field col-12 md:col-6">
+                <AssetEnvironmentSelectField v-model="model.environment"></AssetEnvironmentSelectField>
+            </div>
+            <div class="field col-12 md:col-6">
+                <AssetAccessibleSelectField v-model="model.accessible"></AssetAccessibleSelectField>
+            </div>
+            <div class="field col-12">
+                <label for="technologies">Technologies</label>
+                <TechnologyMultiSelectField v-model="model.technologies"></TechnologyMultiSelectField>
+            </div>
+            <div class="field col-12">
+                <label for="description">Description</label>
+                <MarkdownEditor v-model="model.description"></MarkdownEditor>
+            </div>
+        </div>
+
+        <template #footer>
+            <Button label="Cancel" @click="close" class="p-button-outlined"></Button>
+            <Button label="Save" @click="patch" :loading="loading" icon="pi pi-check" class="p-button-outlined"></Button>
+        </template>
+    </Dialog>
+</template>
